# specs/001-acp-sensitivity-analyzer/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ACP Sensitivity Analyzer API
  description: |
    REST API for ACP (Actual Contribution Percentage) sensitivity analysis.
    Enables programmatic census upload, scenario analysis, and results export
    for mega-backdoor Roth compliance testing.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.example.com/api/v1
    description: Production
...
paths:
...
  /export/{census_id}/csv:
    get:
      tags: [Export]
      summary: Export results as CSV
      description: |
        Export analysis results to CSV format with full audit metadata.
        Includes all input parameters, calculation details, and timestamps.
      operationId: exportCsv
      parameters:
        - $ref: '#/components/parameters/CensusId'
        - name: grid_id
          in: query
          schema:
            type: string
            format: uuid
          description: Export specific grid results
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
          headers:
            Content-Disposition:
              schema:
                type: string
              example: attachment; filename="acp_results_2026-01-12.csv"
        '404':
          description: Census or results not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /export/{census_id}/pdf:
    get:
      tags: [Export]
      summary: Export results as PDF
      description: |
        Export analysis results as a formatted PDF report suitable for
        compliance documentation.
      operationId: exportPdf
      parameters:
        - $ref: '#/components/parameters/CensusId'
        - name: grid_id
          in: query
          schema:
            type: string
            format: uuid
          description: Export specific grid results
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
              example: attachment; filename="acp_report_2026-01-12.pdf"
        '404':
          description: Census or results not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    CensusId:
      name: census_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Census unique identifier
  schemas:
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string